Interfaces, Models as separate projects. Modified CodeIterator and CMCD to make the code modular as intended. First steps toward Naive, AST selectable from command line option.
