+++ CodeDuplicationChecker.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28010.2003
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28729.10
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeDuplicationChecker", "CodeDuplicationChecker\CodeDuplicationChecker.csproj", "{3DC31D7A-B838-41A3-A483-05369B3DECBA}"
 EndProject
@@ -24,8 +24,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UnitTests", "UnitTests", "{
 	ProjectSection(SolutionItems) = preProject
 		SampleCode\UnitTests\RiddledWithDuplicates.cs = SampleCode\UnitTests\RiddledWithDuplicates.cs
 	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NaiveStringComparer", "NaiveStringComparer\NaiveStringComparer.csproj", "{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{8947A8C8-2B78-46E1-BA88-FF3A17E9077C}"
 EndProject
@@ -52,6 +50,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Type4Tests", "Type4Tests", 
 	ProjectSection(SolutionItems) = preProject
 		SampleCode\TypeValidationTests\Type4Tests\VariationsInSyntacticVariants.cs = SampleCode\TypeValidationTests\Type4Tests\VariationsInSyntacticVariants.cs
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Models", "Models\Models.csproj", "{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Interfaces", "Interfaces\Interfaces.csproj", "{6EA320EC-5B67-42A6-909E-C0AC66798EFB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NaiveStringComparer", "NaiveStringComparer\NaiveStringComparer.csproj", "{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -123,18 +127,6 @@ Global
 		{E2FC76E9-F3C6-4FA6-9E45-93EE44B21E09}.Release|x64.Build.0 = Release|Any CPU
 		{E2FC76E9-F3C6-4FA6-9E45-93EE44B21E09}.Release|x86.ActiveCfg = Release|Any CPU
 		{E2FC76E9-F3C6-4FA6-9E45-93EE44B21E09}.Release|x86.Build.0 = Release|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Debug|x64.Build.0 = Debug|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Debug|x86.Build.0 = Debug|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Release|x64.ActiveCfg = Release|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Release|x64.Build.0 = Release|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Release|x86.ActiveCfg = Release|Any CPU
-		{39072DFA-FBC5-4C5D-8B1D-17E11A2394D5}.Release|x86.Build.0 = Release|Any CPU
 		{9C8C1585-25AF-4A10-A285-79B07674E9F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9C8C1585-25AF-4A10-A285-79B07674E9F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9C8C1585-25AF-4A10-A285-79B07674E9F2}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -147,6 +139,42 @@ Global
 		{9C8C1585-25AF-4A10-A285-79B07674E9F2}.Release|x64.Build.0 = Release|Any CPU
 		{9C8C1585-25AF-4A10-A285-79B07674E9F2}.Release|x86.ActiveCfg = Release|Any CPU
 		{9C8C1585-25AF-4A10-A285-79B07674E9F2}.Release|x86.Build.0 = Release|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Debug|x64.Build.0 = Debug|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Debug|x86.Build.0 = Debug|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Release|x64.ActiveCfg = Release|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Release|x64.Build.0 = Release|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Release|x86.ActiveCfg = Release|Any CPU
+		{CA21FD1F-D83F-4A4F-94ED-D05E95F9611E}.Release|x86.Build.0 = Release|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Debug|x64.Build.0 = Debug|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Debug|x86.Build.0 = Debug|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Release|x64.ActiveCfg = Release|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Release|x64.Build.0 = Release|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Release|x86.ActiveCfg = Release|Any CPU
+		{6EA320EC-5B67-42A6-909E-C0AC66798EFB}.Release|x86.Build.0 = Release|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Debug|x64.Build.0 = Debug|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Debug|x86.Build.0 = Debug|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Release|x64.ActiveCfg = Release|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Release|x64.Build.0 = Release|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Release|x86.ActiveCfg = Release|Any CPU
+		{528577F6-D5AD-4EEB-AF65-22CDCC7D076B}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
