+++ CodeDuplicationCheckerIntegrationTests/app.config
