+++ CodeDuplicationCheckerTests/VisualizeDiffsTests.cs
@@ -1,4 +1,4 @@
-﻿using CodeDuplicationChecker;
+﻿using Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
