+++ CountMatrixCloneDetection/Program.cs
@@ -4,8 +4,6 @@ namespace CountMatrixCloneDetection
     {
         public static void Main(string[] args)
         {
-            var cmcdResults = CMCD.Run(@"..\\..\\..\\");
-            return;
         }
     }
 }
