+++ Interfaces/ICodeComparer.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.CodeAnalysis;
+
+namespace Interfaces
+{
+    public interface ICodeComparer
+    {
+        double Compare(SyntaxNode methodNode1, SyntaxNode methodNode2);
+    }
+}
