+++ NaiveStringComparer/packages.config
@@ -8,8 +8,6 @@
   <package id="Microsoft.CodeAnalysis.VisualBasic" version="3.1.0" targetFramework="net472" />
   <package id="Microsoft.CodeAnalysis.VisualBasic.Workspaces" version="3.1.0" targetFramework="net472" />
   <package id="Microsoft.CodeAnalysis.Workspaces.Common" version="3.1.0" targetFramework="net472" />
-  <package id="MSTest.TestAdapter" version="1.4.0" targetFramework="net472" />
-  <package id="MSTest.TestFramework" version="1.4.0" targetFramework="net472" />
   <package id="System.Buffers" version="4.5.0" targetFramework="net472" />
   <package id="System.Collections.Immutable" version="1.5.0" targetFramework="net472" />
   <package id="System.Composition" version="1.2.0" targetFramework="net472" />
